fix: refactor e2e structure
