+++ e2e-tests/endpoints/index.ts
@@ -1,9 +1,9 @@
-import { ChainSpec } from '../types';
+import { IChains } from '../types';
 import { kusamaEndpoints } from './kusama';
 import { polkadotEndpoints } from './polkadot';
 import { westendEndpoints } from './westend';
 
-export const endpoints: Record<ChainSpec, string[][]> = {
+export const endpoints: IChains = {
 	kusama: kusamaEndpoints,
 	polkadot: polkadotEndpoints,
 	westend: westendEndpoints,
