+++ e2e-tests/endpoints/kusama/blocks/6819725.json
