+++ e2e-tests/endpoints/kusama/index.ts
@@ -1,29 +1,29 @@
-import block9253 from './9253.json';
-import block2350438 from './2350438.json';
-import block2684767 from './2684767.json';
-import block2713513 from './2713513.json';
-import block2786879 from './2786879.json';
-import block2863649 from './2863649.json';
-import block3102680 from './3102680.json';
-import block3305067 from './3305067.json';
-import block3492347 from './3492347.json';
-import block3573193 from './3573193.json';
-import block4086065 from './4086065.json';
-import block4178606 from './4178606.json';
-import block4542697 from './4542697.json';
-import block5265416 from './5265416.json';
-import block6065418 from './6065418.json';
-import block6396694 from './6396694.json';
-import block6566865 from './6566865.json';
-import block6819725 from './6819725.json';
-import block7354817 from './7354817.json';
-import block7519631 from './7519631.json';
-import block7673144 from './7673144.json';
-import block7944249 from './7944249.json';
-import block8049096 from './8049096.json'
-import block8113510 from './8113510.json';
+import block9253 from './blocks/9253.json';
+import block2350438 from './blocks/2350438.json';
+import block2684767 from './blocks/2684767.json';
+import block2713513 from './blocks/2713513.json';
+import block2786879 from './blocks/2786879.json';
+import block2863649 from './blocks/2863649.json';
+import block3102680 from './blocks/3102680.json';
+import block3305067 from './blocks/3305067.json';
+import block3492347 from './blocks/3492347.json';
+import block3573193 from './blocks/3573193.json';
+import block4086065 from './blocks/4086065.json';
+import block4178606 from './blocks/4178606.json';
+import block4542697 from './blocks/4542697.json';
+import block5265416 from './blocks/5265416.json';
+import block6065418 from './blocks/6065418.json';
+import block6396694 from './blocks/6396694.json';
+import block6566865 from './blocks/6566865.json';
+import block6819725 from './blocks/6819725.json';
+import block7354817 from './blocks/7354817.json';
+import block7519631 from './blocks/7519631.json';
+import block7673144 from './blocks/7673144.json';
+import block7944249 from './blocks/7944249.json';
+import block8049096 from './blocks/8049096.json'
+import block8113510 from './blocks/8113510.json';
 
-export const kusamaEndpoints = [
+export const kusamaBlockEndpoints = [
 	['/blocks/9253', JSON.stringify(block9253)], //v1020
 	['/blocks/2350438', JSON.stringify(block2350438)], //v1062
 	['/blocks/2684767', JSON.stringify(block2684767)], //v2005
@@ -49,3 +49,7 @@ export const kusamaEndpoints = [
 	['/blocks/8049096', JSON.stringify(block8049096)], //v9050
 	['/blocks/8113510', JSON.stringify(block8113510)], //v9070
 ];
+
+export const kusamaEndpoints = {
+	blocks: kusamaBlockEndpoints
+}
