+++ e2e-tests/endpoints/polkadot/index.ts
@@ -1,16 +1,16 @@
-import block943438 from './943438.json';
-import block1603025 from './1603025.json';
-import block1662525 from './1662525.json';
-import block1993625 from './1993625.json';
-import block2392625 from './2392625.json';
-import block3592619 from './3592619.json';
-import block3892620 from './3892620.json';
-import block4092619 from './4092619.json';
-import block4392619 from './4392619.json';
-import block4947391 from './4947391.json';
-import block5705186 from './5705186.json';
+import block943438 from './blocks/943438.json';
+import block1603025 from './blocks/1603025.json';
+import block1662525 from './blocks/1662525.json';
+import block1993625 from './blocks/1993625.json';
+import block2392625 from './blocks/2392625.json';
+import block3592619 from './blocks/3592619.json';
+import block3892620 from './blocks/3892620.json';
+import block4092619 from './blocks/4092619.json';
+import block4392619 from './blocks/4392619.json';
+import block4947391 from './blocks/4947391.json';
+import block5705186 from './blocks/5705186.json';
 
-export const polkadotEndpoints = [
+const polkadotBlockEndpoints = [
 	['/blocks/943438', JSON.stringify(block943438)], //v17
 	['/blocks/1603025', JSON.stringify(block1603025)], //v18
 	['/blocks/1662525', JSON.stringify(block1662525)], //v23
@@ -23,3 +23,8 @@ export const polkadotEndpoints = [
 	['/blocks/4947391', JSON.stringify(block4947391)], //v30
 	['/blocks/5705186', JSON.stringify(block5705186)], //v9050
 ];
+
+
+export const polkadotEndpoints = {
+	blocks: polkadotBlockEndpoints
+}
