+++ e2e-tests/endpoints/westend/index.ts
@@ -1,14 +1,14 @@
-import block3032259 from './3032259.json';
-import block3813629 from './3813629.json';
-import block3889918 from './3889918.json';
-import block4781573 from './4781573.json';
-import block5277929 from './5277929.json';
-import block5480769 from './5480769.json';
-import block5493461 from './5493461.json';
-import block5495855 from './5495855.json';
-import block5657482 from './5657482.json';
+import block3032259 from './blocks/3032259.json';
+import block3813629 from './blocks/3813629.json';
+import block3889918 from './blocks/3889918.json';
+import block4781573 from './blocks/4781573.json';
+import block5277929 from './blocks/5277929.json';
+import block5480769 from './blocks/5480769.json';
+import block5493461 from './blocks/5493461.json';
+import block5495855 from './blocks/5495855.json';
+import block5657482 from './blocks/5657482.json';
 
-export const westendEndpoints = [
+export const westendBlockEndpoints = [
 	['/blocks/3032259', JSON.stringify(block3032259)], //v45
 	['/blocks/3813629', JSON.stringify(block3813629)], //v45
 	['/blocks/3889918', JSON.stringify(block3889918)], //v45
@@ -19,3 +19,7 @@ export const westendEndpoints = [
 	['/blocks/5495855', JSON.stringify(block5495855)], //v9000
 	['/blocks/5657482', JSON.stringify(block5657482)], //v9010
 ];
+
+export const westendEndpoints = {
+	blocks: westendBlockEndpoints
+}
